Tabs as their own component
